Move interfaces to -util, add healthcheck to HBaseConnection
