+++ stroom-stats-hbase-client/src/main/java/stroom/stats/hbase/connection/HBaseConnection.java
@@ -21,6 +21,7 @@
 
 package stroom.stats.hbase.connection;
 
+import com.codahale.metrics.health.HealthCheck;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.apache.hadoop.hbase.TableName;
@@ -33,6 +34,7 @@ import org.slf4j.LoggerFactory;
 import stroom.stats.hbase.HBaseStatisticConstants;
 import stroom.stats.hbase.exception.HBaseException;
 import stroom.stats.properties.StroomPropertyService;
+import stroom.stats.util.healthchecks.HasHealthCheck;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -46,8 +48,9 @@ import java.time.Instant;
  * A dependency on this class will result in a connection being made to Zookeeper
  */
 @Singleton
-public class HBaseConnection {
+public class HBaseConnection implements HasHealthCheck{
     private final Configuration configuration;
+    private final String quorum;
     private final boolean autoCreateTables;
     // private final HTablePool pool;
 
@@ -68,7 +71,7 @@ public class HBaseConnection {
     @Inject
     public HBaseConnection(final StroomPropertyService propertyService) {
 
-        final String quorum = propertyService.getPropertyOrThrow(
+        quorum = propertyService.getPropertyOrThrow(
                 HBaseStatisticConstants.HBASE_ZOOKEEPER_QUORUM_PROPERTY_NAME);
 
         LOGGER.info("Initialising HBaseTableConfiguration to quorum: {}", quorum);
@@ -149,6 +152,7 @@ public class HBaseConnection {
     public HBaseConnection(final Connection connection) {
         this.sharedClusterConnection = connection;
         this.configuration = connection.getConfiguration();
+        this.quorum = connection.getConfiguration().get(HBASE_ZOOKEEPER_QUORUM_PROPERTY_NAME);
         autoCreateTables = true;
     }
 
@@ -184,5 +188,26 @@ public class HBaseConnection {
         } catch (final IOException e) {
             throw new HBaseException("Unable close HBase connection", e);
         }
+    }
+
+    @Override
+    public HealthCheck.Result getHealth() {
+        try {
+            HBaseAdmin.checkHBaseAvailable(configuration);
+
+            return HealthCheck.Result.builder()
+                    .healthy()
+                    .withMessage("HBase running on quorum " + quorum)
+                    .build();
+        } catch (Exception e) {
+            return HealthCheck.Result.builder()
+                    .unhealthy(e)
+                    .build();
+        }
+    }
+
+    @Override
+    public String getName() {
+        return "HBaseConnection";
     }
 }
