+++ stroom-stats-service/src/main/java/stroom/stats/streams/StatisticsFlatMappingService.java
@@ -23,10 +23,10 @@ import com.codahale.metrics.health.HealthCheck;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import stroom.stats.api.StatisticType;
-import stroom.stats.mixins.HasHealthCheck;
-import stroom.stats.mixins.HasRunState;
-import stroom.stats.mixins.Startable;
-import stroom.stats.mixins.Stoppable;
+import stroom.stats.util.healthchecks.HasHealthCheck;
+import stroom.stats.util.HasRunState;
+import stroom.stats.util.Startable;
+import stroom.stats.util.Stoppable;
 import stroom.stats.properties.StroomPropertyService;
 import stroom.stats.streams.mapping.CountStatToAggregateFlatMapper;
 import stroom.stats.streams.mapping.ValueStatToAggregateFlatMapper;
