+++ stroom-stats-util/src/main/java/stroom/stats/util/HasName.java
@@ -17,7 +17,7 @@
  * along with Stroom-Stats.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package stroom.stats.mixins;
+package stroom.stats.util;
 
 public interface HasName {
